+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Updated frontend to display version
+  [#59](https://github.com/chrisleekr/binance-trading-bot/issues/59)
 - Support monitoring multiple coins simultaneously
   [#77](https://github.com/chrisleekr/binance-trading-bot/pull/77)
 - Added `docker-stack.yml` for the Portainer -